+++ tests/test_export_avro_idl.py
@@ -16,9 +16,7 @@ from datacontract.lint.resolve import resolve_data_contract_from_location
 
 
 def test_ir():
-    contract = resolve_data_contract_from_location(
-        "fixtures/lint/valid_datacontract_references.yaml", inline_definitions=True
-    )
+    contract = resolve_data_contract_from_location("fixtures/lint/valid_datacontract_ref.yaml", inline_definitions=True)
     expected = AvroIDLProtocol(
         name="OrdersLatest",
         description="Successful customer orders in the webshop.\n"
@@ -43,9 +41,7 @@ def test_ir():
 
 
 def test_avro_idl_str():
-    contract = resolve_data_contract_from_location(
-        "fixtures/lint/valid_datacontract_references.yaml", inline_definitions=True
-    )
+    contract = resolve_data_contract_from_location("fixtures/lint/valid_datacontract_ref.yaml", inline_definitions=True)
     expected = dedent(
         """
           /** Successful customer orders in the webshop.
@@ -66,9 +62,7 @@ def test_avro_idl_str():
 
 def test_avro_idl_cli_export():
     runner = CliRunner()
-    result = runner.invoke(
-        app, ["export", "./fixtures/lint/valid_datacontract_references.yaml", "--format", "avro-idl"]
-    )
+    result = runner.invoke(app, ["export", "./fixtures/lint/valid_datacontract_ref.yaml", "--format", "avro-idl"])
     if result.exit_code:
         print(result.output)
     assert result.exit_code == 0