+++ CHANGELOG.md
@@ -12,6 +12,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- datacontract lint: Linter 'Field references existing field' too many values to unpack (expected
+  2) (#586)
+
 ## [0.10.18] - 2024-01-18
 
 ### Fixed